Organized files into three packages, test, view, and main. This establishes the MVC pattern where MC is currently in main, and V is in view. All unit tests are in the test package. This commit also updated the .gitignore to include machine and IDE specific files when using eclipse.
