--- src/module-info.java
@@ -1,10 +1,0 @@
-/**
- * 
- */
-/**
- * @author Bryan
- *
- */
-module Hockey {
-	requires java.desktop;
-}
